backend/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lifehub-backend"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "fastapi",
    "sqlalchemy",
    "pydantic",
    "mariadb==1.1.9",
    "python-dotenv",
    "python-jose",
    "argon2-cffi",
    "pytz",
    "uvicorn",
    "python-multipart"
]

[tool.setuptools.packages]
find = {}

[project.scripts]
lifehub-api = "lifehub.app.api:run"
lifehub-fetch = "lifehub.app.fetch:run"
lifehub-setup = "lifehub.app.setup:setup"
lifehub-clean = "lifehub.app.setup:clean"

[tool.ruff]
lint.extend-select = ["I"]

[tool.mypy]
strict = true

exclude = '^tests/'
